Single Responsability Fix and re factoring
