+++ server.ts
@@ -1,13 +1,10 @@
-import express from 'express';
 import dotenv from 'dotenv';
 import connectToMongo from './Infrastructure/database/mongo.js';
 import logger from './Infrastructure/logging/logger.js';
-import api from './API/index.js';
 import config from './config.js'
+import startAPIServer from './API/index';
 
 dotenv.config();
-const app = express();
-app.use(express.json());
 
 logger.info('Starting server...');
 
@@ -21,7 +18,4 @@ configPropiedades.forEach(propiedad => {
 
 await connectToMongo();
 
-app.use('/api', api);
-
-const PORT = process.env.PORT || 3000;
-app.listen(PORT, () => logger.info(`Server running on port ${PORT}`));
+startAPIServer();
